rpmsg: resmgr: stub regulator requests when not defined
The remote requests for regulator API have been stubbed out
in the case when regulator framework is not included in the
kernel. The standard regulator api stubs return success when
CONFIG_REGULATOR is not defined, but returning this success
status to remote processors can make the requestors to think
that the regulator requests have been satisfied, when they
are not in reality.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>