omap: remoteproc: implement suspend and resume callbacks
Implement resume and suspend callbacks, that way omap remote
processors can support suspend and resume.

The host processor can not save the context of the remote processors
for that reason it sends messages to the remote processor requesting
enter to suspend. The remote processor can ack the request or cancel
the request depending of if it is still doing some work. However
when there is a system suspend the remote processor should suspend
even if it is doing some job. Therefore, in the system suspend case
we send a different suspend request to the remote processor saying
that the suspend is forced and it must enter to suspend.
Also, if the remoteproc has a register that can say if the remote
processor is in idle state, it can be specify in the platform data that
way the host can know if it is in idle state without the need to send
a message to the remote processor.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>