[sru-workflow-manager] Task search refactored out of the process_bugs method
The process_bugs method is big and complicated. I'm trying to reduce
it's complications by extracting out bits and pieces.

This change pulls out the code that calls LP to do a task search into
it's own funciton. That function just produces a list of the bug ids
that match the search criteria. The logic to handle a single bug is
also removed. Now that process_bugs processes a single bug (given it's
bug id) we don't have to do a task search when we just want to work
on a few, specified bugs.

process_bugs has been renamed to shank.

Signed-off-by: Brad Figg <brad.figg@canonical.com>